+++ src/main/java/com/tristankechlo/toolleveling/utils/ButtonHelper.java
@@ -20,14 +20,14 @@ public final class ButtonHelper {
             return ToolLevelingConfig.allowLevelingUselessEnchantments.getValue();
         } else if (entry.status == ButtonStatus.BREAK) {
             return ToolLevelingConfig.allowLevelingBreakingEnchantments.getValue();
-        } else if (entry.status == ButtonStatus.BLACKLISTED || entry.status == ButtonStatus.CAPPED || entry.status == ButtonStatus.MAXLEVEL) {
+        } else if (entry.status == ButtonStatus.BLACKLISTED || entry.status == ButtonStatus.CAPPED
+                || entry.status == ButtonStatus.MAX_LEVEL || entry.status == ButtonStatus.MIN_LEVEL) {
             return false;
         }
         return false;
     }
 
-    public static ButtonEntry getButtonEntry(ToolLevelingTableHandledScreen parent, Enchantment enchantment,
-                                             int level) {
+    public static ButtonEntry getButtonEntry(ToolLevelingTableHandledScreen parent, Enchantment enchantment, int level) {
         List<Enchantment> whitelist = ToolLevelingConfig.whitelist.getValue();
         List<Enchantment> blacklist = ToolLevelingConfig.blacklist.getValue();
         ButtonEntry buttonEntry = new ButtonEntry(parent, enchantment, level);
@@ -41,6 +41,10 @@ public final class ButtonHelper {
         else if (whitelist.isEmpty() && blacklist.contains(enchantment)) {
             buttonEntry.status = ButtonStatus.BLACKLISTED;
         }
+        // check if the enchantment is over the set minimum level
+        else if (!Utils.isEnchantmentOverMinimum(enchantment, level)) {
+            buttonEntry.status = ButtonStatus.MIN_LEVEL;
+        }
         // check if the enchantment is allowed to level up
         // determinated by the config enchantmentCaps
         else if (Utils.isEnchantmentAtCap(enchantment, level)) {
@@ -49,7 +53,7 @@ public final class ButtonHelper {
         // although the level is defined as an integer, the actual maximum is a short
         // a higher enchantment level than a short will result in a negative level
         else if (level >= Short.MAX_VALUE) {
-            buttonEntry.status = ButtonStatus.MAXLEVEL;
+            buttonEntry.status = ButtonStatus.MAX_LEVEL;
         }
         // leveling these enchantments will do absolutely nothing
         else if (enchantment.getMaxLevel() == 1) {
@@ -68,7 +72,6 @@ public final class ButtonHelper {
         return Text.translatable(entry.name).formatted(getButtonTextFormatting(entry));
     }
 
-    @SuppressWarnings("resource")
     public static List<Text> getButtonToolTips(ButtonEntry data) {
         List<Text> tooltip = new ArrayList<>();
         tooltip.add(Text.translatable(data.name).formatted(Formatting.AQUA));
@@ -86,7 +89,6 @@ public final class ButtonHelper {
         return tooltip;
     }
 
-    @SuppressWarnings("resource")
     public static Formatting getButtonTextFormatting(ButtonEntry entry) {
         Formatting format = Formatting.RESET;
         if (Utils.freeCreativeUpgrades(MinecraftClient.getInstance().player)) {
@@ -102,11 +104,12 @@ public final class ButtonHelper {
     }
 
     public enum ButtonStatus {
-        NORMAL,
-        BLACKLISTED,
-        USELESS,
-        MAXLEVEL,
-        BREAK,
-        CAPPED;
+        NORMAL, // nothing special, can be leveled
+        BLACKLISTED, // enchantment is blacklisted, or not on the whitelist
+        USELESS, // leveling this enchantment will have no effect
+        BREAK, // enchantment will break when leveled higher
+        MAX_LEVEL, // enchantment is at the possible maximum level (Short.MAX_VALUE)
+        CAPPED, // enchantment is at the maximum level (enchantmentCaps)
+        MIN_LEVEL // enchantment is not over the set minimum level (minimumEnchantmentLevel)
     }
 }